pkg/netcat: move SSLOptions.GenerateTLSConfiguration() to temporary source
In a subsequent patch, I'll rewrite SSLOptions.GenerateTLSConfiguration().
So that hunks of the old and the new function not be interleaved in that
patch, move the current code verbatim to a new source file, temporarily.

Best viewed with "git show --no-renames --color-moved=zebra".
