+++ pkg/netcat/ssl.go
@@ -3,13 +3,6 @@
 // license that can be found in the LICENSE file.
 package netcat
 
-import (
-	"crypto/tls"
-	"crypto/x509"
-	"fmt"
-	"os"
-)
-
 type SSLOptions struct {
 	// In connect mode, this option transparently negotiates an SSL session
 	// In server mode, this option listens for incoming SSL connections
@@ -26,44 +19,4 @@ type SSLOptions struct {
 	Ciphers []string
 	SNI     string   // (Server Name Indication) Tell the server the name of the logical server Ncat is contacting
 	ALPN    []string // List of protocols to send via the Application-Layer Protocol Negotiation
-}
-
-func (s *SSLOptions) GenerateTLSConfiguration() (*tls.Config, error) {
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: !s.VerifyTrust,
-	}
-
-	if s.CertFilePath == "" || s.KeyFilePath == "" {
-		return nil, fmt.Errorf("both  certificate and key file must be provided")
-	}
-
-	cer, err := tls.LoadX509KeyPair(s.CertFilePath, s.KeyFilePath)
-	if err != nil {
-		return nil, fmt.Errorf("connection: %w", err)
-	}
-
-	tlsConfig.Certificates = []tls.Certificate{cer}
-
-	if s.VerifyTrust {
-		caCert, err := os.ReadFile(s.TrustFilePath)
-		if err != nil {
-			return nil, fmt.Errorf("cannot read CA certificate: %w", err)
-		}
-		caCertPool := x509.NewCertPool()
-		if !caCertPool.AppendCertsFromPEM(caCert) {
-			return nil, fmt.Errorf("cannot append CA certificate to pool")
-		}
-
-		tlsConfig.RootCAs = caCertPool
-	}
-
-	if s.SNI != "" {
-		tlsConfig.ServerName = s.SNI
-	}
-
-	if s.ALPN != nil {
-		tlsConfig.NextProtos = s.ALPN
-	}
-
-	return tlsConfig, nil
 }
